// src/app/pages/school-performance/types/visit-plan-member-task.ts
export interface VisitPlanMemberTask {
  id?: number;
  createdBy?: number;
  memberNameAr?: string;
  memberNameEn?: string;
  role?: string;
  visitPlanMemberTaskAssignmentId?: number;
  taskSlot?: string;
  taskType?: string;
  domain?: string;
  subject?: string;
  customSubject?: string;
  place?: string;
  targetAudience?: string;
  customTargetAudience?: string;
  grade?: string;
  teacherNameAr?: string;
  teacherNameEn?: string;
  activity?: string;
  customActivity?: string;
  notes?: string;
}